Move useWindowDimensions to hooks util
